requirements.txt
Flask
opencv-python
dlib
spotipy
numpy
pandas
scikit-learn
python-dotenv
imutils
Flask-SocketIO
keras
eventlet
screeninfo

# Used for the KAN ML Model
brainflow
PyQt5
pykan
pyqtgraph
torch
PyYAML
tqdm